Add Python 3 interpreter
A simple but effective python 3 interpreter with a purposefully concise but arcane implementation is included. A set of anonymous functions will provide state transitions for the "machine" state. A simple parser cleans up and prepare a list of instructions that will be then fed to such functions for making the machine do the job.

Finally, included in the same file, the ability of reading files using the first argument of direct invocation is implemented.